Se quitó la reactividad a los casos de uso que hacía falta. Fueron eliminados los casos de uso de eliminar y mostrar. Ha sido incorporado los servicios de buscar Movimientos por el bolsilloId, y de buscar Bolsillos por el UsuarioId. Ha sido añadido la funcionalidad de que el saldoDisponible d eun Bolsillo aumente o disminuya dependiendo del Movimiento que se cree. Ahora se puede ver los movimientos cuando se busca un bolsillo, y cada vez que se elimine un bolsillo, también se eliminan los movimientos

+++ Proyecto-finanzas-back/Finanzas/src/main/java/co/com/finanzas/useCase/TransformacionBolsilloUseCase.java
@@ -2,6 +2,8 @@ package co.com.finanzas.useCase;
 
 import co.com.finanzas.domain.infra.repository.BolsilloData;
 import co.com.finanzas.domain.infra.repository.IBolsilloDataRepository;
+import co.com.finanzas.domain.infra.repository.IMovimientoDataRepository;
+import co.com.finanzas.domain.infra.repository.MovimientoData;
 import co.com.finanzas.domain.model.bolsillo.Bolsillo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,13 +14,15 @@ public class TransformacionBolsilloUseCase {
     @Autowired
     private IBolsilloDataRepository data;
 
-    public BolsilloData transformar(Bolsillo bolsillo){
-        BolsilloData bolsilloData = new BolsilloData(bolsillo.getIdPro(),bolsillo.getNombre().value(), bolsillo.getSaldoDisponible().value(),bolsillo.getUid().value(),bolsillo.getEsAhorro().value(),bolsillo.getPorcentajeAhorro().value());
-        return bolsilloData;
-    }
+    @Autowired
+    private TransformacionMovimientoUseCase transformacionMovimientoUseCase;
 
     public Iterable<BolsilloData> listar(){
         return data.findAll();
+    }
+
+    public Iterable<BolsilloData> listarPorUid(String uid){
+        return data.findByUid(uid);
     }
 
     public BolsilloData listarPorId(String id){
@@ -26,11 +30,19 @@ public class TransformacionBolsilloUseCase {
     }
 
     public String eliminar(String id){
-        try{
-            data.deleteById(id);
-            return  "Se eliminó el bolsillo con éxito";
-        } catch (Exception e){
-            return "No fue posible eliminar el bolsillo";
+        BolsilloData bolsillo = listarPorId(id);
+        if (bolsillo.getSaldoDisponible() == 0){
+            try{
+                Iterable<MovimientoData> movimientos = transformacionMovimientoUseCase.listarPorBolsilloId(id);
+                transformacionMovimientoUseCase.eliminarMovimientosPorBolsillo(movimientos);
+                data.deleteById(id);
+                return  "Se eliminó el bolsillo con éxito";
+            } catch (Exception e){
+                return "No fue posible eliminar el bolsillo";
+            }
+        }else {
+            return "No puedes eliminar un bolsillo si tienes saldo disponible. Por favor retire su saldo del bolsillo";
         }
+
     }
 }
